Fix: P2-5 — add explicit np.float32 dtype conversion in VectorStore
  upsert_kp_vector() and upsert_topic_vector() now call vector.astype(np.float32)
  before tobytes(). Guarantees consistent dtype on write regardless of caller.
  Read path already uses np.frombuffer(..., dtype=np.float32) — now
  write-then-read is safe even if caller passes float64 or other dtype.

  Deferred for the rest (P0-1, P0-2, P1-3, P1-4, P2-6, P2-7, P2-8):
  All documented with rationale in AgentChatRoom/feedback.md.

+++ exam_analyzer/src/stores/vector_store.py
@@ -19,7 +19,7 @@ class VectorStore:
                    (kp_id, vector, adjustment_count, updated_at)
                    VALUES (?, ?, COALESCE((SELECT adjustment_count FROM kp_vectors
                     WHERE kp_id=?), 0) + 1, datetime('now'))""",
-                (kp_id, vector.tobytes(), kp_id),
+                (kp_id, vector.astype(np.float32).tobytes(), kp_id),
             )
             self._mgr.maybe_commit()
 
@@ -54,7 +54,7 @@ class VectorStore:
                 """INSERT OR REPLACE INTO topic_vectors
                    (topic_id, vector, member_kp_count, updated_at)
                    VALUES (?, ?, ?, datetime('now'))""",
-                (topic_id, vector.tobytes(), member_count),
+                (topic_id, vector.astype(np.float32).tobytes(), member_count),
             )
             self._mgr.maybe_commit()
